Changed the documentation style to a better one
The new documentation style is very attractive for most code editors.
If comments are written in this format for every functions, while calling those functions your IDE (supported IDEs like VS code and Jet brain IDEs) will show the developer more information about the function in a popup. While calling these kind of documented functions the Code editor helps  the developers to use correct amount of arguments and while writing each of the arguments in the function it'll show which datatype to use and a short info about the individual arguments.

Additionally writing author's name in each functions' documentation will help a team of developers to identify each function's author on the go while coding.
While updating an existing function you could give two asterisk instead of one for the old @version tags and put the new version in another new line with only one asterisk at the beginning to take that value as the new version number.